docker/requirements.txt
Flask
prometheus-flask-exporter
flask-compress
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-urllib
redis
requests
flask_zipkin
